# docs/swagger.yaml
basePath: /
definitions:
  http.CreatePRRequest:
    properties:
      author_id:
        description: '@Example: user123'
        example: user123
        type: string
      pull_request_id:
        description: '@Example: pr-123'
        example: pr-123
        type: string
      pull_request_name:
        description: '@Example: "Add new feature"'
        example: Add new feature
        type: string
    type: object
  http.CreateTeamRequest:
    properties:
      members:
        items:
          $ref: '#/definitions/schemas.User'
        type: array
      name:
        description: '@Example: backend'
        example: backend
        type: string
    type: object
  http.LoginRequest:
    properties:
      password:
        description: '@Example: admin'
        example: admin
        type: string
      user_id:
        description: '@Example: admin'
        example: admin
        type: string
    type: object
  http.LoginResponse:
    properties:
      role:
        description: '@Example: admin'
        example: admin
        type: string
      token:
        description: '@Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  http.MergePRRequest:
    properties:
      pull_request_id:
        description: '@Example: pr-123'
        example: pr-123
        type: string
    type: object
  http.ReassignPRRequest:
    properties:
      old_user_id:
        description: '@Example: user123'
        example: user123
        type: string
      pull_request_id:
        description: '@Example: pr-123'
        example: pr-123
        type: string
    type: object
  http.SetUserActiveRequest:
    properties:
      is_active:
        description: '@Example: true'
        example: true
        type: boolean
      user_id:
        description: '@Example: user123'
        example: user123
        type: string
    type: object
  schemas.Team:
    properties:
      members:
        items:
          $ref: '#/definitions/schemas.User'
        type: array
      team_name:
        type: string
    type: object
  schemas.User:
    properties:
      is_active:
        type: boolean
      team_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for assigning code reviewers to pull requests
  title: Review Assigner API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Логин для получения JWT
      tags:
      - Auth
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - Health
  /pullRequest/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: PR data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreatePRRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать PR
      tags:
      - PullRequests
  /pullRequest/merge:
    post:
      consumes:
      - application/json
      parameters:
      - description: PR ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.MergePRRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Merge PR
      tags:
      - PullRequests
  /pullRequest/reassign:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reassign data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ReassignPRRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reassign reviewer
      tags:
      - PullRequests
  /stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить статистику назначений
      tags:
      - Statistics
  /team/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Team data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать команду
      tags:
      - Teams
  /team/get:
    get:
      parameters:
      - description: Team name
        example: '"backend"'
        in: query
        name: team_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Team'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить команду
      tags:
      - Teams
  /users/getReview:
    get:
      parameters:
      - description: User ID
        example: '"user123"'
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить PR пользователя
      tags:
      - Users
  /users/setIsActive:
    post:
      consumes:
      - application/json
      parameters:
      - description: User active data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.SetUserActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Установить активность пользователя
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"