# internal/domain/schemas/user.go
package schemas

type User struct {
    ID       string `json:"user_id" db:"user_id"`
    Username string `json:"username" db:"username"`
    TeamName string `json:"team_name" db:"team_name"`
    IsActive bool   `json:"is_active" db:"is_active"`
}